Move thunk lowering code into its own file.
This required lifting some shared code into the iterator.h header.
